+++ web/example/mount/gin/main.go
@@ -1,43 +1,19 @@
-package main
+package gin
 
 import (
-	"sync"
+	_ "github.com/Rhaqim/buckt/web" // Web package for side effects
 
-	// _ "github.com/lib/pq"
-
-	"fmt"
 	"log"
 	"strings"
 
 	"github.com/Rhaqim/buckt"
-	_ "github.com/Rhaqim/buckt/web"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
 
-	var err error
-	// var db *sql.DB
-
-	// // Postgres database
-	// conn_string := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable",
-	// 	"localhost", 5432, "postgres", "password", "postgres")
-
-	// db, err = sql.Open("postgres", conn_string)
-	// if err != nil {
-	// 	log.Fatalf("Failed to connect to the database: %v", err)
-	// }
-
-	// File cache
-	cache := NewCache()
-
 	// Initialize Buckt
 	opts := buckt.BucktConfig{
-		// DB: buckt.DBConfig{
-		// 	Driver:   buckt.Postgres,
-		// 	Database: db,
-		// }, // Pass the database connection
-		Cache: cache,
 		Log: buckt.LogConfig{
 			LogTerminal: false,
 			LogFile:     "logs",
@@ -98,50 +74,4 @@ func main() {
 
 	// Start the server
 	r.Run(":8080")
-}
-
-type Cache struct {
-	// Cache
-	mu    sync.RWMutex
-	store map[string]any
-}
-
-func NewCache() *Cache {
-	return &Cache{
-		store: make(map[string]any),
-	}
-}
-
-func (c *Cache) GetBucktValue(key string) (any, error) {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-
-	fmt.Println("Cache get", key)
-
-	if val, ok := c.store[key]; ok {
-
-		fmt.Println("Cache hit", key)
-
-		return val, nil
-	}
-
-	return nil, nil
-}
-
-func (c *Cache) SetBucktValue(key string, value any) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	fmt.Println("Cache set", key)
-
-	c.store[key] = value
-	return nil
-}
-
-func (c *Cache) DeleteBucktValue(key string) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	delete(c.store, key)
-	return nil
 }
